Refactor header image fields into separate php partial
